chore: Search VC VM Coordinator 삭제
